COMMERCE-5334 dispatch - extract Helper to api module
