+++ modules/apps/dispatch/dispatch-api/src/main/java/com/liferay/dispatch/executor/DispatchTaskExecutorHelper.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.dispatch.executor;
+
+import java.util.Set;
+
+/**
+ * @author Matija Petanjek
+ */
+public interface DispatchTaskExecutorHelper {
+
+	public DispatchTaskExecutor getDispatchTaskExecutor(
+		String taskExecutorType);
+
+	public String getDispatchTaskExecutorName(String taskExecutorType);
+
+	public Set<String> getDispatchTaskExecutorTypes();
+
+}
